refactor: Component cleanup
