+++ src/renderer/src/components/catalog/EditFileDialog.tsx
@@ -1,6 +1,4 @@
 import { Button } from '@/components/ui/button'
-import { Input } from '@/components/ui/input'
-import { Label } from '@/components/ui/label'
 import {
   Dialog,
   DialogContent,
@@ -8,22 +6,11 @@ import {
   DialogDescription,
   DialogFooter
 } from '@/components/ui/dialog'
-import { Checkbox } from '@/components/ui/checkbox'
-import {
-  Select,
-  SelectContent,
-  SelectItem,
-  SelectTrigger,
-  SelectValue
-} from '@/components/ui/select'
-import { Pencil, Check, FolderOpen, X } from 'lucide-react'
+import { Pencil, Check } from 'lucide-react'
 import type { IModFile } from '@shared/schema'
-import { SIDECAR_EXPLANATION } from '@/lib/catalog/types'
 import type { EditFormState } from '@/lib/catalog/types'
-import { CATEGORIES } from '@shared/categories'
 import type { RequiredModsActions } from '@/lib/catalog/useRequiredModsActions'
-import { RequiredModsEditor } from '@/components/catalog/RequiredModsEditor'
-import { InfoTooltip } from '@/components/ui/info-tooltip'
+import { ModFileFormFields } from '@/components/catalog/ModFileFormFields'
 
 interface EditFileDialogProps {
   open: boolean
@@ -70,120 +57,17 @@ export function EditFileDialog({
         </div>
 
         <div className="space-y-4 p-4 overflow-y-auto">
-          <div className="space-y-2">
-            <Label htmlFor="edit-name">Name</Label>
-            <Input
-              id="edit-name"
-              value={form.name}
-              onChange={(e) => setForm((prev) => ({ ...prev, name: e.target.value }))}
-              className="bg-app-secondary border-app"
-            />
-          </div>
-
-          <div className="space-y-2">
-            <Label htmlFor="edit-version">Version</Label>
-            <Input
-              id="edit-version"
-              value={form.version}
-              onChange={(e) => setForm((prev) => ({ ...prev, version: e.target.value }))}
-              className="bg-app-secondary border-app"
-            />
-          </div>
-
-          <div className="space-y-2">
-            <Label htmlFor="edit-url">URL</Label>
-            <Input
-              id="edit-url"
-              value={form.url}
-              onChange={(e) => setForm((prev) => ({ ...prev, url: e.target.value }))}
-              className="bg-app-secondary border-app"
-            />
-          </div>
-
-          <div className="space-y-2">
-            <Label>Category</Label>
-            <Select
-              value={form.category}
-              onValueChange={(val) => setForm((prev) => ({ ...prev, category: val }))}
-            >
-              <SelectTrigger className="bg-app-secondary border-app">
-                <SelectValue placeholder="Uncategorized" />
-              </SelectTrigger>
-              <SelectContent>
-                {CATEGORIES.map((cat) => (
-                  <SelectItem key={cat} value={cat}>
-                    {cat.replace(/_/g, ' ').replace(/\b\w/g, (l) => l.toUpperCase())}
-                  </SelectItem>
-                ))}
-              </SelectContent>
-            </Select>
-          </div>
-
-          {/* Config Template Section */}
-          <div className="space-y-2">
-            <Label>Config Template (optional)</Label>
-            <p className="text-xs text-app-muted">
-              This config will seed new protocols that include this mod file.
-            </p>
-            <div className="flex gap-2">
-              <Input
-                value={form.configTemplate?.filePath || ''}
-                readOnly
-                placeholder="No config linked"
-                className="bg-app-secondary border-app flex-1 text-xs"
-              />
-              {form.configTemplate ? (
-                <Button
-                  type="button"
-                  variant="outline"
-                  onClick={onClearConfigFile}
-                  className="bg-app-secondary border-app"
-                  title="Remove config template"
-                >
-                  <X className="h-4 w-4" />
-                </Button>
-              ) : null}
-              <Button
-                type="button"
-                variant="outline"
-                onClick={onBrowseConfigFile}
-                className="bg-app-secondary border-app"
-              >
-                <FolderOpen className="h-4 w-4" />
-              </Button>
-            </div>
-            {form.configTemplate ? (
-              <p className="text-xs text-green-500 font-mono">
-                MD5: {form.configTemplate.md5Hash.slice(0, 16)}...
-              </p>
-            ) : null}
-          </div>
-
-          <RequiredModsEditor
-            loadOrder={form.loadOrder}
+          <ModFileFormFields
+            idPrefix="edit"
+            form={form}
+            setForm={setForm}
             selectableFiles={selectableFiles}
-            onAddFromCatalog={requiredModsActions.handleAddFromCatalog}
-            onBrowseFile={requiredModsActions.handleBrowseFile}
-            onRemove={requiredModsActions.handleRemove}
-            onMoveUp={requiredModsActions.handleMoveUp}
-            onMoveDown={requiredModsActions.handleMoveDown}
-            onToggleSidecar={requiredModsActions.handleToggleSidecar}
-            onNameChange={requiredModsActions.handleNameChange}
+            requiredModsActions={requiredModsActions}
+            onBrowseConfigFile={onBrowseConfigFile}
+            onClearConfigFile={onClearConfigFile}
+            configTemplateHelpText="This config will seed new protocols that include this mod file."
+            configTemplatePlaceholder="No config linked"
           />
-
-          <div className="flex items-center gap-2">
-            <Checkbox
-              id="edit-sidecar"
-              checked={form.sidecarOnly}
-              onCheckedChange={(checked) =>
-                setForm((prev) => ({ ...prev, sidecarOnly: checked === true }))
-              }
-            />
-            <Label htmlFor="edit-sidecar" className="text-sm font-normal">
-              Sidecar mod
-            </Label>
-            <InfoTooltip text={SIDECAR_EXPLANATION} />
-          </div>
 
           {selectedFile?.hashValue && (
             <div className="text-xs text-app-muted">
